Reorganize composable into settings package.
