+++ app/src/main/java/com/gallopdevs/athanhelper/ui/settings/NotificationsOption.kt
@@ -1,0 +1,44 @@
+package com.gallopdevs.athanhelper.ui.settings
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Switch
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.gallopdevs.athanhelper.R
+import com.gallopdevs.athanhelper.ui.theme.AthanHelperTheme
+
+@Composable
+fun NotificationsOption(checked: Boolean, onCheckedChange: (Boolean) -> Unit) {
+    Row(
+        verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.SpaceBetween,
+        modifier = Modifier.fillMaxWidth()
+    ) {
+        SettingsLabel(
+            settingsIcon = painterResource(id = R.drawable.bell_icon),
+            settingsHeader = stringResource(id = R.string.notifications)
+        )
+        Switch(
+            checked = checked,
+            onCheckedChange = onCheckedChange,
+            modifier = Modifier
+                .padding(end = 30.dp)
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun NotificationsOptionPreview() {
+    AthanHelperTheme {
+        NotificationsOption(checked = false, onCheckedChange = {})
+    }
+}
